✨ add profile card
